chore: reflect tcp only in name and reduce code duplication
