--- src/Testcontainers/Configurations/WaitStrategies/UntilTcpConnected.cs
@@ -1,43 +1,0 @@
-using System.Net.Sockets;
-using System.Threading.Tasks;
-using DotNet.Testcontainers.Containers;
-
-namespace DotNet.Testcontainers.Configurations
-{
-    /// <summary>
-    /// Wait for a TCP connection to be established.
-    /// </summary>
-    /// <remarks>
-    /// Some docker configurations (DockerForMac & Rancher on Mac & Windows) allow establishing a connection to any mapped port.
-    /// https://forums.docker.com/t/port-forwarding-of-exposed-ports-behaves-unexpectedly/15807
-    /// </remarks>
-    internal sealed class UntilTcpConnected(int port) : IWaitUntil
-    {
-        private readonly int _containerPort = port;
-
-        public async Task<bool> UntilAsync(IContainer container)
-        {
-            ushort hostPort;
-            try
-            {
-                hostPort = container.GetMappedPublicPort(_containerPort);
-            }
-            catch
-            {
-                return false;
-            }
-
-            using var tcpClient = new TcpClient();
-            try
-            {
-                await tcpClient.ConnectAsync(container.Hostname, hostPort)
-                    .ConfigureAwait(false);
-                return tcpClient.Connected;
-            }
-            catch
-            {
-                return false;
-            }
-        }
-    }
-}
