+++ tests/Testcontainers.Platform.Windows.Tests/WindowsContainerTest.cs
@@ -64,14 +64,14 @@ public abstract class WindowsContainerTest : IAsyncLifetime
     }
 
     [UsedImplicitly]
-    public sealed class UntilHostPortIsAvailable : WindowsContainerTest
+    public sealed class UntilHostTcpPortIsAvailable : WindowsContainerTest
     {
-        public UntilHostPortIsAvailable()
+        public UntilHostTcpPortIsAvailable()
             : base(new ContainerBuilder()
                 .WithImage(CommonImages.ServerCore)
                 .WithEntrypoint("PowerShell", "-NoLogo", "-Command")
                 .WithCommand("$tcpListener = [System.Net.Sockets.TcpListener]80; $tcpListener.Start(); Start-Sleep -Seconds 120")
-                .WithWaitStrategy(Wait.ForWindowsContainer().UntilHostPortAvailable(80))
+                .WithWaitStrategy(Wait.ForWindowsContainer().UntilHostTcpPortAvailable(80))
                 .Build())
         {
         }
