+++ tests/Testcontainers.Tests/Unit/Configurations/HostTcpPortWaitStrategyTest.cs
@@ -8,7 +8,7 @@ namespace DotNet.Testcontainers.Tests.Unit
   using JetBrains.Annotations;
   using Xunit;
 
-  public sealed class HostPortWaitStrategyTest : IAsyncLifetime
+  public sealed class HostTcpPortWaitStrategyTest : IAsyncLifetime
   {
     private const int listeningPort = 49152;
     private const int mappedPort = 49153;
@@ -20,7 +20,7 @@ namespace DotNet.Testcontainers.Tests.Unit
       .WithCommand("EXEC:cat")
       .WithPortBinding(listeningPort, true)
       .WithPortBinding(mappedPort, true)
-      .WithWaitStrategy(Wait.ForUnixContainer().UntilHostPortAvailable(listeningPort))
+      .WithWaitStrategy(Wait.ForUnixContainer().UntilHostTcpPortAvailable(listeningPort))
       .Build();
 
     public async ValueTask InitializeAsync()
@@ -35,9 +35,9 @@ namespace DotNet.Testcontainers.Tests.Unit
     }
 
     [Fact]
-    public async Task hostPortWaitStrategyIsSucceeded()
+    public async Task HostTcpPortWaitStrategyIsSucceeded()
     {
-      var hostPortWaitStrategy = new HostPortWaitStrategy(listeningPort);
+      var hostPortWaitStrategy = new HostTcpPortWaitStrategy(listeningPort);
 
       var succeeded = await hostPortWaitStrategy.UntilAsync(_container);
 
@@ -49,9 +49,9 @@ namespace DotNet.Testcontainers.Tests.Unit
     /// This test might fail in docker configurations where mapped ports are always listened eg. DockerForMac (https://forums.docker.com/t/port-forwarding-of-exposed-ports-behaves-unexpectedly/15807/2)
     /// </summary>
     [Fact]
-    public async Task hostPortWaitStrategyFailsWhenPortNotListening()
+    public async Task HostTcpPortWaitStrategyFailsWhenPortNotListening()
     {
-      var hostPortWaitStrategy = new HostPortWaitStrategy(mappedPort);
+      var hostPortWaitStrategy = new HostTcpPortWaitStrategy(mappedPort);
 
       var succeeded = await hostPortWaitStrategy.UntilAsync(_container);
 
@@ -59,9 +59,9 @@ namespace DotNet.Testcontainers.Tests.Unit
     }
 
     [Fact]
-    public async Task hostPortWaitStrategyFailsWhenPortNotMapped()
+    public async Task HostTcpPortWaitStrategyFailsWhenPortNotMapped()
     {
-      var hostPortWaitStrategy = new HostPortWaitStrategy(unmappedPort);
+      var hostPortWaitStrategy = new HostTcpPortWaitStrategy(unmappedPort);
 
       var succeeded = await hostPortWaitStrategy.UntilAsync(_container);
